+++ tests/test_server.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 from paxos.server import Server
-from paxos.core import Message
+from paxos.core import Message, ProposalNumber
 from paxos.protocol import ProposalNumber
